requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
tensorboard==2.13.0
matplotlib==3.7.1
seaborn==0.12.2
pandas==2.0.3
numpy==1.25.0
Pillow==10.0.0
wordcloud==1.9.2
scikit-learn==1.3.0
nltk==3.8.1
jsonlines==3.1.0
gensim==4.3.1
tqdm==4.65.0